+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changelog
 0.1.2 (upcoming)
 ----------------
 
+* Added logaugment.set which calls reset() and add() in order.
 * Make logaugment.add work for all collections.Mapping instances.
 
 0.1.1 (September 8, 2014)